# API/static/schema/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Social Media App",
    "description": "This is a API for Flight Ticket\n\n\n               \n               \n               \n               \n",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000/"
    },
    {
      "url": "https://flightx-4fxd.onrender.com/"
    }
  ],
...
  "paths": {
    "/userSignUp": {
      "post": {
        "description": "Register a user in the database.The username should be unique. The username must be atleast 6 characters and atmost 20 characters. The username must contain only alphabets. The email must be valid. The password must contain only alphabets and numbers and must be of atleast 6 characters long and atmost of 20 characters long.",
        "tags": [
          "User"
        ],
        "requestBody": {
          "description": "A json object containing username,email,and a password. Username and email must be unique.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/userDetails"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully a user is created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/operationSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorReport"
                }
              }
            }
          }
        }
      }
    },
...
    "/userLogin": {
      "post": {
        "description": "Logins a user with username or email and password.Generates an API key.",
        "tags": [
          "User"
        ],
        "requestBody": {
          "description": "A json object containg username (or email) and password.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/loginDetails"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully a user is logged in and an API key is created. If the user is already logged in then the API key is returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/loginSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Login Fails when password is wrong or user name doesn't exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorReport"
                }
              }
            }
          }
        }
      }
    },
    "/searchFlight": {
      "get": {
        "description": "Allows user to get the list of flights according to the fliter. The filter options is passed to in the URL as query.If no query is given then all flights are returned irrespective of the date inorder to give user. The list of flights are sorted according to their date.",
        "tags": [
          "User"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "flight_name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "flight_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "query",
            "description": "Must be in YYYY-MM-DD",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Flights reults are returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/flights"
                }
              }
            }
          },
          "400": {
            "description": "Either the api key is invalid or no flights are found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorReport"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "key"
      }
    },
    "schemas": {
      "userDetails": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "loginDetails": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "loginSuccess": {
        "type": "object",
        "properties": {
          "Message": {
            "type": "string"
          },
          "key": {
            "type": "string"
          }
        }
      },
...
      "flights": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer"
          },
          "Flights": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "flight_id": {
                  "type": "string"
                },
                "flight_name": {
                  "type": "string"
                },
                "date": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "operationSuccess": {
        "type": "object",
        "properties": {
          "Message": {
            "type": "string"
          }
        }
      },
      "errorReport": {
        "type": "object",
        "properties": {
          "Error": {
            "type": "string"
          }
        }
      },
...
    }
  }
}